chatbot/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# AI Services
openai==1.54.0  # Optional
google-generativeai==0.3.2  # Gemini (FREE)

# Vector Database (Qdrant)
qdrant-client==1.12.0

# PostgreSQL Database (Neon)
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Authentication (Optional - Bonus Feature)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
httpx==0.27.2
tenacity==9.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
